// src/app/models/delivery-item.ts
export class DeliveryItem {
  id: number = null;
  delivery_method: string = null;
  portal_publication_id: number = null;
  dose: number = null;
  frequency_id: number = null;
  frequency:any = null;
  publication:any = null;
  stock_date:string = null;
  quantity:number = null;
  each: number = null;
  duration: string = null;
  more_info: any = null;
  delivery_day: any = null;
  from_date: string = null;
  covered_by_insurance: boolean = null;
  insurance_auth_number: string = null;
  insurance_auth: {
    auth_autogenerated_date: string, auth_code: string, auth_expiry_date: string,
    auth_issue_date: string, current_recurrence: number, id: number, configured: boolean, recurrence_period: number, total_recurrence: number
  } = null;
}